fixed snake render issue

+++ Linked-List-Snake/source/Level/LevelModel.cpp
@@ -6,10 +6,11 @@ namespace Level
 
 	LevelModel::~LevelModel() = default;
 
-	void LevelModel::initialize(float width, float height)
+	void LevelModel::initialize(int width, int height)
 	{
-		cell_width = width / number_of_columns;
-		cell_height = height / number_of_rows;
+		cell_width = static_cast<float>(width) / static_cast<float>(number_of_columns);
+		cell_height = static_cast<float>(height) / static_cast<float>(number_of_rows);
+
 	}
 
 	float LevelModel::getCellWidth()
